Refactor: create a `SharedProjectPage` component.
- The same structure is shared between `[username]/[projectName]` and
`[username]/[porjectName]/[multiProjectName]`.
